Create dummy Git repository container
Create a dummy Git repository which can be used for generating demo
videos.

A custom shell prompt showing Git repository status has been added to
improve user experience when making changes.

It was necessary to prevent `vim` from fixing end of line as the demo
Git repository did not have all files ending with a new line character.
This meant changes to these files also changed the last character
causing unexpected differences. From a demo perspective this was not
useful.

A login shell was also required to allow the `profile.d` directory from
having all scripts executed. It was ignored if this argument was not
set.

A Docker compose file was also created to simplify starting this
container.