Add do the change slides
Add slides for the do the change section.

Multiple changes are represented as stacked commits and are rebased
against master to unstack them. This may not always be possible if
there is a dependency on a lower layer commit. To reduce complexity,
this explanation has not been mentioned.

It is important to also mention that the interactive rebase code block
has been modified to show only the frequently used commands. The commit
hash shown for each command has been removed and replaced with the
branch name for simplicity.