Add prepare the commit slides
Add slides for the prepare the commit section.

It is necessary to mention that the background image does not load for
the slides that have quotes. This is a limitation in how layouts work
(not all layouts support a background image). A custom layout will be
created at a later date to use the provided background image.