Add Gitea container
When using the Docker image to create screenshots or record videos it
was sometimes necessary to have a remote repository accessible.

A dependency was added to the Docker compose file to start the Gitea
service.

This allows a remote repository to be set to `server:3000` and perform
all remote operations on this endpoint.

A user must be created in Gitea first otherwise authentication will
fail.