Add icon set dependencies
Icon sets added from Iconify are insalled as packages. When Slidev
encounters the use of these icons, it adds them as development
dependencies and updates the lock file.

By doing this manually, this negates the need for Slidev to do this
automatically and keeps the repository in sync with the running state.