Split demo video tape files
The application VHS is used to record the demo videos. A new feature
was added that allowed for the tape files to be sourced from other
files. This allowed for a more modular approach to build the videos.

As part of the change to the modular components, some of the tape logic
was modified. This was especially apparent for hidden sections where
the process had significantly shorter wait times. As these components
are now shared, they need to use the original visible wait times. This
slows down the video creation.

It was not possible to add the video dimensions to the init tape as
this introduced a bug with incorrect margins. Other settings were able
to be migrated to this file.

As all recording is done within a Docker container, the shell was
unnecessarily set and has been removed.