Simplify restoring files
When using the Git restore command, by default it only restores to the
worktree. When providing the staged argument, this changes this
functionality to only restore to the staging area and keeps the
worktree the same.

It is preferrable to restore to both the worktree and staging area
however this increases the complexity as both arguments must be
provided. Simplifying this to no argument at all restores to just the
worktree.

An extra Git command is required to stage this change however this does
not add much complexity.