Update demo video text editor
When applying Git commits, Vi was previously used. Changing to Vim
allowed for the use of a plugin supporting Git commits which provided
many benefits including adding colour.

The plugin did not hard wrap commit messages to 72 characters and this
option was necessary to add to the custom settings. The previous method
for setting options did not handle multiple options. Using a
configuration file was necessary.