Simplify process for new commits
The previous example for what to do when new commits were added while
still working on a branch was based on stacked branches. This was too
complicated for the target audence.

The simplified example is based on a single branch.

A new feature was added to Mermaid when using Git Graph diagrams. This
allows for parallel commits which makes it easier to transition from
one state to another while retaining the absolute position of previous
commits. In this example, it was useful to show how the rebase moved
the branch commit forward. This was also applied for a previous example
where it was beneficial.